Add new methods to services, add pagination methods

+++ src/tg_bot_float_db_app/database/models/subscription_model.py
@@ -16,7 +16,7 @@ class SubscriptionModel(Base):
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
 
-    user_id: Mapped[BigInteger] = mapped_column(
+    user_id: Mapped[int] = mapped_column(
         BigInteger, ForeignKey("user.id", ondelete="cascade"), nullable=False
     )
